requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.8.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
langfuse==2.60.3
openai==1.75.0
packaging==24.2
pluggy==1.5.0
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
setuptools==78.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
wrapt==1.17.2
pinecone-client>=3.0.0
